// src/app/classes/app-state.interface.ts
export interface ListingViewState {
  order: Array<string | number>;
  page?: LookupTable<Array<number | string>>;
  total?: number;
  query?: Query;
  loading: LookupTable<boolean>;
}

export interface ResultEntry {
  [prop: string]: any;
  response: APIResponse;
}

export interface ModelState<T = any> {
  byId: LookupTable<T>;
  loading: LookupTable<boolean>;
  results: LookupTable<ResultEntry>;
}
// ...
export interface LookupTable<T> {
  [id: string]: T;
}